refactor: overlay provider ergonomics
