+++ package/src/state-store/index.ts
@@ -1,3 +1,4 @@
 export * from './audio-player';
 export * from './in-app-notifications-store';
 export * from './audio-player-pool';
+export * from './message-overlay-store';
